# Conflicts:
#	.gitignore
#	CMakeLists.txt
#	plugins/logicalaccess/plugins/cards/epass/utils.cpp
#	plugins/pluginscards/mifare/mifarecommands.cpp
#	plugins/pluginsreaderproviders/iso7816/commands/desfireev1iso7816commands.cpp
#	plugins/pluginsreaderproviders/pcsc/pcscreaderunit.cpp
#	src/cards/samkeystorage.cpp
#	src/dynlibrary/librarymanager.cpp
#	src/settings.cpp

+++ src/services/accesscontrol/formats/bithelper.cpp
@@ -63,7 +63,7 @@ BitsetStream BitHelper::revert(const BitsetStream &data, unsigned int dataLength
 
             BitsetStream linedData;
 
-            linedData.concat(tmp, 0, tmp.size() * 8);
+            linedData.concat(tmp, 0, static_cast<unsigned int>(tmp.size()) * 8);
 
             return linedData;
         }
